Moving Button into common components
