+++ js/containers/report/feedback-panel.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react' // eslint-disable-line
 import ReactDom from 'react-dom'
 
-import Button from '../../components/report/button'
+import Button from '../../components/common/button'
 import FeedbackFilter from '../../components/report/feedback-filter'
 import FeedbackOverview from '../../components/report/feedback-overview'
 import FeedbackRow from '../../components/report/feedback-row'
